Fix logging in now that 32-bit support disappeared
In the March 2024 the 32-bit client was removed, so we need to modify
some of the boot targets that no longer exist. The DirectX 9 client is
gone too, but that will be dealt with in a later commit.

(cherry picked from commit 516c0926a67d86a502d8b99ee02803cd9adf3314)