Don't try to update Wine/DXVK if we're using a custom version
In the event that GitHub is down, the game is actually unplayable as the
launcher doesn't allow you to skip updating these components (for now.)
To prevent this, you can use a custom wine you have installed but the
launcher was still erroneously updating it. Now it won't try to update
DXVK/Wine if you don't request it to, although I'm not sure about tying
DXVK to the wine option right now.