When displaying default location for install/wine prefix, hide the path
For most users, this ends up being a long useless string (because of the
UUID included.) If it's the default, Astra-managed path then it simply
says "Default Location". When it's a custom path, it shows the full path
as per usual.

This affects the visible game paths, but also the wine prefix path.