Overhaul asset updating for Wine and DXVK, make URLs configurable
This used to be just be a hardcoded URL (I know, I know...) but now it
fetches the latest release from each repository. To make it even more
future-proof, the URLs of the repositories and even the GitHub API used
to fetch them is configurable.

UI to expose that will come in the next commit.