Make Astra more useful when offline
When you're offline, all of the asset update checks fail (of course) and
thus can prevent you from logging in. While it's possible to work around
this already by toggling certain things off such as Dalamud, the
built-in Wine etc - we should still be able to play if we have all of
that downloaded.

Now there is buttons in the network fail prompts that allow you to
continue if you understand the risks. This would also make Astra more
resilient to issues like GitHub being temporarily unavailable.