Make our filesystem permissions more explicit
Fixes #37, but in a normal setup we only normally need to access the
pictures directory and the Steam compatibility tools directory. If you
do use an existing game install (not controlled by us) then you have to
give us permission for that directory, but that's a Flatpak problem in
general.